+++ tests/units/test_rp_agent.py
@@ -8,6 +8,7 @@ from behave.model_core import Status
 from delayed_assert import assert_expectations, expect
 from prettytable import PrettyTable
 from reportportal_client.client import RPClient
+from reportportal_client.logs.log_manager import MAX_LOG_BATCH_PAYLOAD_SIZE
 
 from behave_reportportal.behave_agent import BehaveAgent, create_rp_service
 from behave_reportportal.config import Config, LogLayout
@@ -175,6 +176,8 @@ def test_create_rp_service_init(mock_rps):
                 launch_id=None,
                 retries=None,
                 mode="DEFAULT",
+                log_batch_size=20,
+                log_batch_payload_size=MAX_LOG_BATCH_PAYLOAD_SIZE
             )
         ],
         any_order=True,
@@ -308,6 +311,7 @@ def verify_start_feature(mock_feature, config):
     mock_feature.description = ["A", "B"]
     ba = BehaveAgent(config, mock_rps)
     ba.start_feature(mock_context, mock_feature, some_key="some_value")
+    # noinspection PyProtectedMember
     mock_rps.start_test_item.assert_called_once_with(
         name="feature_name",
         start_time=123,
@@ -375,6 +379,7 @@ def verify_start_scenario(mock_scenario, config):
     ba = BehaveAgent(config, mock_rps)
     ba._feature_id = "feature_id"
     ba.start_scenario(mock_context, mock_scenario, some_key="some_value")
+    # noinspection PyProtectedMember
     mock_rps.start_test_item.assert_called_once_with(
         name="scenario_name",
         start_time=123,
@@ -387,6 +392,7 @@ def verify_start_scenario(mock_scenario, config):
         test_case_id=ba._test_case_id(mock_scenario),
         some_key="some_value",
     )
+    # noinspection PyProtectedMember
     assert ba._scenario_id == "scenario_id", (
         f"Invalid scenario_id:\nActual: {ba._scenario_id}\n"
         f"Expected: {'scenario_id'}\n"
@@ -633,7 +639,7 @@ def test_finish_failed_step_scenario_based(mock_timestamp, config):
 
 
 @mock.patch("behave_reportportal.behave_agent.timestamp")
-def test_log_exception_without_message(mock_timestamp):
+def test_log_exception_without_message(mock_timestamp, config):
     mock_timestamp.return_value = 123
     mock_step = mock.Mock()
     mock_step.exception = None
@@ -653,7 +659,7 @@ def test_log_exception_without_message(mock_timestamp):
 
 def test_rp_is_none():
     ba = BehaveAgent(Config(), None)
-    ba.start_step(mock.Mock(), mock.Mock)
+    ba.start_step(mock.Mock(), mock.Mock())
     assert ba._step_id is None