+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Client version updated on [5.3.0](https://github.com/reportportal/client-Python/releases/tag/5.3.0), by @HardNorth
+- move debug mode to RPClient init, by @nicke46
+- Make agent not finish the launch if a launch_id was supplied at start, by @nicke46
 
 ## [2.0.1]
 ### Fixed