+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Stub files for `config` and `behave_agent` modules, by @HardNorth
+- `log_batch_size` and `log_batch_payload_size` config parameters, by @HardNorth
+### Changed
+- Build script updates, by @HardNorth
+### Fixed
+- Launch attributes parsing, by @HardNorth
 
 ## [2.0.2]
 ### Changed