+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Stub files for `config` and `behave_agent` modules, by @HardNorth
+- `log_batch_size` and `log_batch_payload_size` config parameters, by @HardNorth
+### Changed
+- Build script updates, by @HardNorth
+### Fixed
+- Launch attributes parsing, by @HardNorth
+
+## [2.0.2]
 ### Changed
 - Client version updated on [5.3.0](https://github.com/reportportal/client-Python/releases/tag/5.3.0), by @HardNorth
 - move debug mode to RPClient init, by @nicke46